requirements.txt
requests~=2.24.0
flask~=1.1.2
python-dotenv~=0.13.0
pycryptodomex
flask-cors